refactor: Move `serialize_str_into` to `utils.rs`
