+++ src/nbt/tag.rs
@@ -44,17 +44,6 @@ impl NbtTag {
         self.serialize_data_into(bytes);
     }
 
-    pub fn serialize_str_into(s: &str, bytes: &mut BytesMut) {
-        if s.is_empty() {
-            bytes.put_u16(0);
-            return;
-        }
-
-        let java_string = simd_cesu8::encode(s);
-        bytes.put_u16(java_string.len() as u16);
-        bytes.put_slice(&java_string);
-    }
-
     pub fn serialize_data(&self) -> Bytes {
         let mut bytes = BytesMut::new();
         self.serialize_data_into(&mut bytes);
@@ -74,7 +63,7 @@ impl NbtTag {
                 bytes.put_i32(byte_array.len() as i32);
                 bytes.put_slice(byte_array);
             }
-            NbtTag::String(string) => Self::serialize_str_into(string, bytes),
+            NbtTag::String(string) => serialize_str_into(string, bytes),
             NbtTag::List(list) => {
                 bytes.put_u8(list.first().unwrap_or(&NbtTag::End).get_type_id());
                 bytes.put_i32(list.len() as i32);
