+++ src/nbt/utils.rs
@@ -1,7 +1,7 @@
 use std::fmt::{self, Display, Formatter};
 
 use crate::error::Error;
-use bytes::Buf;
+use bytes::{Buf, BufMut, BytesMut};
 use simd_cesu8::decode;
 
 pub const END_ID: u8 = 0;
@@ -23,6 +23,17 @@ pub fn get_nbt_string(bytes: &mut impl Buf) -> Result<String, Error> {
     let string_bytes = bytes.copy_to_bytes(len);
     let string = decode(&string_bytes).map_err(|_| Error::InvalidJavaString)?;
     Ok(string.to_string())
+}
+
+pub fn serialize_str_into(s: &str, bytes: &mut BytesMut) {
+    if s.is_empty() {
+        bytes.put_u16(0);
+        return;
+    }
+
+    let java_string = simd_cesu8::encode(s);
+    bytes.put_u16(java_string.len() as u16);
+    bytes.put_slice(&java_string);
 }
 
 // This can be improved once rust-lang/rust#132980 is resolved:
