+++ src/serde/ser.rs
@@ -1,7 +1,6 @@
 use crate::error::Error::UnsupportedType;
 use crate::error::{Error, Result};
 use crate::nbt::utils::*;
-use crate::NbtTag;
 use bytes::{BufMut, BytesMut};
 use crab_nbt::nbt::utils::END_ID;
 use serde::ser::Impossible;
@@ -32,7 +31,7 @@ impl Serializer {
         match &mut self.state {
             State::Named(name) | State::Array { name, .. } => {
                 self.output.put_u8(tag);
-                NbtTag::serialize_str_into(&name, &mut self.output);
+                serialize_str_into(name, &mut self.output);
             }
             State::FirstListElement { len } => {
                 self.output.put_u8(tag);
@@ -177,7 +176,7 @@ impl ser::Serializer for &mut Serializer {
             return Ok(());
         }
 
-        NbtTag::serialize_str_into(&v, &mut self.output);
+        serialize_str_into(v, &mut self.output);
         Ok(())
     }
 
@@ -330,11 +329,11 @@ impl ser::Serializer for &mut Serializer {
         match &mut self.state {
             State::Root(root_name) => {
                 if let Some(root_name) = root_name {
-                    NbtTag::serialize_str_into(&root_name, &mut self.output);
+                    serialize_str_into(root_name, &mut self.output);
                 }
             }
             State::Named(string) => {
-                NbtTag::serialize_str_into(&string, &mut self.output);
+                serialize_str_into(string, &mut self.output);
             }
             State::FirstListElement { len } => {
                 self.output.put_i32(*len);
